* 'master' of github.com:Jumpscale/jumpscale_core8: (483 commits)
  Rewrite j.do.execute to run commands using asyncio.subprocess and coroutines (#491)
  added lunx to the docs, \n \n as it was removed from the cygwin default installation and now requires manual install (#490)
  remove sslVerify=False (#471)
  automated the installation of apt-cyg
  bug fix in syncthing , new build method
  less-than-minor-minor
  Revert " bug fix in install tools piping" (#484)
  support for double quotes in hrd added (#452)
  bug fix in install tools piping (#479)
  init only services for the first run (#478)
  Small typo
  Fix link to AYS template repos documentation
  Adding "AYS File Locations & Details" to TOC
  init support for ays, jsdocker autocompletion via click magic variables (#477)
  Reviewed changes of Christophe
  Minor fix (#476)
  Revert "use findAYSRepos instead of repos" (#475)
  use findAYSRepos instead of repos (#474)
  Remove eve (#472)
  Use path and not service key name during getService (#470)
  ...

# Conflicts:
#	lib/JumpScale/sal/fs/SystemFSWalker.py

+++ lib/JumpScale/tools/cuisine/CuisineInstallerDevelop.py
@@ -10,8 +10,8 @@ class actionrun(ActionDecorator):
         ActionDecorator.__init__(self,*args,**kwargs)
         self.selfobjCode="cuisine=j.tools.cuisine.getFromId('$id');selfobj=cuisine.installerdevelop"
 
-
-class CuisineInstallerDevelop:
+base=j.tools.cuisine.getBaseClass()
+class CuisineInstallerDevelop(base):
 
     def __init__(self,executor,cuisine):
         self.executor=executor
@@ -31,6 +31,8 @@ class CuisineInstallerDevelop:
             """
         else:
             C = """
+            python3
+            postgresql
             libpython3.4-dev
             python3.4-dev
             libpython3.5-dev
@@ -39,6 +41,7 @@ class CuisineInstallerDevelop:
         self.cuisine.package.multiInstall(C)
 
         C="""
+        autoconf
         libffi-dev
         gcc
         make
@@ -48,6 +51,7 @@ class CuisineInstallerDevelop:
         pkg-config
         libpq-dev
         libsqlite3-dev
+        vim
         #net-tools
         """
         self.cuisine.package.multiInstall(C)
@@ -56,6 +60,9 @@ class CuisineInstallerDevelop:
     def pip(self):
         self.cuisine.installer.base()
         self.python()
+        if self.cuisine.core.isMac:
+            return
+
         C="""
             #important remove olf pkg_resources, will conflict with new pip
             rm -rf /usr/lib/python3/dist-packages/pkg_resources
@@ -76,22 +83,7 @@ class CuisineInstallerDevelop:
 
         self.python()
         self.pip(action=True)
-
-        if not self.cuisine.core.isArch:
-            #install brotli
-            C="""
-            cd $tmpDir/
-            sudo rm -rf brotli/
-            git clone https://github.com/google/brotli.git
-            cd $tmpDir/brotli/
-            python setup.py install
-            cd tools
-            make
-            cp $tmpDir/brotli/tools/bro /usr/local/bin/
-            rm -rf $tmpDir/brotli
-            """
-            C=self.cuisine.core.args_replace(C)
-            self.cuisine.core.run_script(C,force=False)
+        self.brotli()
 
         #python etcd
         C="""
@@ -110,7 +102,8 @@ class CuisineInstallerDevelop:
         self.cuisine.core.run_script(C,force=False)
 
         C="""
-        cffi==1.5.2
+        # cffi==1.5.2
+        cffi
         paramiko
 
         msgpack-python
@@ -122,6 +115,7 @@ class CuisineInstallerDevelop:
 
         certifi
         docker-py
+        http://carey.geek.nz/code/python-fcrypt/fcrypt-1.3.1.tar.gz
 
         gitlab3
         gitpython
@@ -173,7 +167,13 @@ class CuisineInstallerDevelop:
         python-telegram-bot
         colorlog
         path.py
+        dnspython3
+        packet-python
+        gspread
+        oauth2client
         """
+        if not self.cuisine.core.isCygwin:
+            C += "pillow"
         self.cuisine.pip.multiInstall(C,upgrade=True)
 
 
@@ -184,18 +184,15 @@ class CuisineInstallerDevelop:
             """
             self.cuisine.pip.multiInstall(C,upgrade=True)
 
-        self.cuisine.apps.redis.build()
+        if  not self.cuisine.core.isCygwin:
+            self.cuisine.apps.redis.build()
 
-        """
-        install dnspython3
-        """
-        self.dnspython3()
 
     @actionrun(action=True)
     def jumpscale8(self):
         if self.cuisine.installer.jumpscale_installed():
             return
-        self.installJS8Deps()
+        self.installJS8Deps(force=False)
 
         if self.cuisine.core.isUbuntu or self.cuisine.core.isArch:
 
@@ -207,21 +204,53 @@ class CuisineInstallerDevelop:
             C=self.cuisine.core.args_replace(C)
             self.cuisine.core.run(C)
         elif self.cuisine.core.isMac:
-            cmd = """sudo mkdir -p /opt
-            # sudo chown -R despiegk:root /opt
-            ruby -e \"$ (curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)\""""
+            cmd = "export TMPDIR=~/tmp;mkdir -p $TMPDIR;cd $TMPDIR;rm -f install.sh;curl -k https://raw.githubusercontent.com/Jumpscale/jumpscale_core8/master/install/install.sh > install.sh;bash install.sh"
             self.cuisine.core.run(cmd)
         else:
             raise j.exceptions.RuntimeError("platform not supported yet")
 
+    @actionrun(action=True)
+    def cleanup(self):
+        self.cuisine.core.run("apt-get clean")
+        self.cuisine.core.dir_remove("/var/tmp/*")
+        self.cuisine.core.dir_remove("/etc/dpkg/dpkg.cfg.d/02apt-speedup")
+        self.cuisine.core.dir_remove("$tmpDir")
+        self.cuisine.core.dir_ensure("$tmpDir")
+
+
 
     @actionrun(action=True)
-    def dnspython3(self):
-        C = """
-            cd $tmpDir
-            wget http://www.dnspython.org/kits3/1.12.0/dnspython3-1.12.0.tar.gz
-            tar -xf dnspython3-1.12.0.tar.gz
-            cd dnspython3-1.12.0
-            ./setup.py install
-            """
+    def brotli(self):
+        C="""
+        cd /tmp
+        sudo rm -rf brotli/
+        git clone https://github.com/google/brotli.git
+        cd /tmp/brotli/
+        ./configure
+        make bro
+        cp /tmp/brotli/bin/bro /usr/local/bin/
+        rm -rf /tmp/brotli
+        """
+        C=self.cuisine.core.args_replace(C)
+        self.cuisine.core.run_script(C,force=True)
+
+    @actionrun()
+    def xrdp(self):
+        """
+        builds a full xrdp, this can take a while
+        """
+        C="""
+        cd /root
+        git clone https://github.com/scarygliders/X11RDP-o-Matic.git
+        cd X11RDP-o-Matic
+        bash X11rdp-o-matic.sh
+        ln -fs /usr/bin/Xvfb /etc/X11/X
+        apt-get update
+        apt-get install  -y --force-yes lxde lxtask
+        echo 'pgrep -U $(id -u) lxsession | grep -v ^$_LXSESSION_PID | xargs --no-run-if-empty kill' > /bin/lxcleanup.sh
+        chmod +x /bin/lxcleanup.sh
+        echo '@lxcleanup.sh' >> /etc/xdg/lxsession/LXDE/autostart
+        echo '#!/bin/sh -xe\nrm -rf /tmp/* /var/run/xrdp/* && service xrdp start && startx' > /bin/rdp.sh
+        chmod +x /bin/rdp.sh
+        """
         self.cuisine.core.run_script(C)
